Navigation scrolling functionality achieved
When the window is at the top position, the background color of the navigation bar is set to null. When the window is at any other scrolling position, the background is set to the light gray color with an opacity of 95%. Also, when the user scrolls the navigation bar stays fixed at the top.