feat: About Us page is fully functional
About Us has been completed with exception of minor styling. It includes unstyled buttons that select pictures and paragraphs describing the history of the company in Lorem Ipsum text. The next specific update of this page will include animations for both pics and paragraphs.

Also included in this commit are two new javascript files to successfully split the JS load of the different pages.