Add functionality to carousel on contact page
I completed the code needed to make the carousel work. This required a big change to the class names in the HTML for both the Services and Contacts pages. Making this change made the Javascript components function for both pages. With the functionality done, the full content is visible on the page, so the last fine tuning of the CSS styling can be done.